requirements.txt
# AI & Transcription
google-genai
faster-whisper==1.0.0

# PyTorch with CUDA 12.1
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# Audio & Speech
pyaudio
sounddevice
simpleaudio
pydub
edge_tts
soundfile

# Input / Output
pynput
opencv-python
Pillow
emoji

# Utilities
python-dotenv
numpy